requirements.txt
django-cors-headers
djangorestframework
pillow
django
pandas
